+++ tests/Services/LoggerTest.php
@@ -9,7 +9,7 @@ use ImageConverterWebP\Services\Logger;
 /**
  * @covers \ImageConverterWebP\Core\Converter::__construct
  * @covers \ImageConverterWebP\Services\Logger::__construct
- * @covers \ImageConverterWebP\Services\Logger::add_webp_meta_to_attachment
+ * @covers \ImageConverterWebP\Services\Logger::add_logs_for_webp_conversions
  * @covers icfw_get_settings
  */
 class LoggerTest extends TestCase {
@@ -25,7 +25,7 @@ class LoggerTest extends TestCase {
 		\WP_Mock::tearDown();
 	}
 
-	public function test_add_webp_meta_to_attachment_does_not_log_error_if_log_option_is_not_enabled() {
+	public function test_add_logs_for_webp_conversions_does_not_log_error_if_log_option_is_not_enabled() {
 		$webp = Mockery::mock( '\WP_Error' )->makePartial();
 
 		$options = [
@@ -37,17 +37,12 @@ class LoggerTest extends TestCase {
 			->with( 'icfw', [] )
 			->andReturn( $options );
 
-		\WP_Mock::userFunction( 'is_wp_error' )
-			->once()
-			->with( $webp )
-			->andReturn( true );
-
-		$this->logger->add_webp_meta_to_attachment( $webp, 1 );
+		$this->logger->add_logs_for_webp_conversions( $webp, 1 );
 
 		$this->assertConditionsMet();
 	}
 
-	public function test_add_webp_meta_to_attachment_logs_error_if_wp_error_is_true() {
+	public function test_add_logs_for_webp_conversions_logs_error_if_wp_error_is_true() {
 		$webp = Mockery::mock( '\WP_Error' )->makePartial();
 
 		$webp->shouldReceive( 'get_error_message' )
@@ -64,7 +59,7 @@ class LoggerTest extends TestCase {
 			->andReturn( $options );
 
 		\WP_Mock::userFunction( 'is_wp_error' )
-			->twice()
+			->once()
 			->with( $webp )
 			->andReturn( true );
 
@@ -80,39 +75,7 @@ class LoggerTest extends TestCase {
 			)
 			->andReturn( 100 );
 
-		$this->logger->add_webp_meta_to_attachment( $webp, 1 );
-
-		$this->assertConditionsMet();
-	}
-
-	public function test_add_webp_meta_to_attachment_updates_post_meta() {
-		$webp = 'https://example.com/wp-content/uploads/2024/01/sample.webp';
-
-		$options = [
-			'logs' => true,
-		];
-
-		\WP_Mock::userFunction( 'get_option' )
-			->once()
-			->with( 'icfw', [] )
-			->andReturn( $options );
-
-		\WP_Mock::userFunction( 'is_wp_error' )
-			->twice()
-			->with( $webp )
-			->andReturn( false );
-
-		\WP_Mock::userFunction( 'get_post_meta' )
-			->once()
-			->with( 1, 'icfw_img', true )
-			->andReturn( '' );
-
-		\WP_Mock::userFunction( 'update_post_meta' )
-			->once()
-			->with( 1, 'icfw_img', 'https://example.com/wp-content/uploads/2024/01/sample.webp' )
-			->andReturn( null );
-
-		$this->logger->add_webp_meta_to_attachment( $webp, 1 );
+		$this->logger->add_logs_for_webp_conversions( $webp, 1 );
 
 		$this->assertConditionsMet();
 	}
