+++ tests/Services/MetaDataTest.php
@@ -1,0 +1,54 @@
+<?php
+
+namespace ImageConverterWebP\Tests\Services;
+
+use Mockery;
+use WP_Mock\Tools\TestCase;
+use ImageConverterWebP\Services\MetaData;
+
+/**
+ * @covers \ImageConverterWebP\Core\Converter::__construct
+ * @covers \ImageConverterWebP\Services\MetaData::__construct
+ * @covers \ImageConverterWebP\Services\MetaData::add_webp_meta_to_attachment
+ * @covers icfw_get_settings
+ */
+class MetaDataTest extends TestCase {
+	public MetaData $metadata;
+
+	public function setUp(): void {
+		\WP_Mock::setUp();
+
+		$this->metadata = new MetaData();
+	}
+
+	public function tearDown(): void {
+		\WP_Mock::tearDown();
+	}
+
+	public function test_add_webp_meta_to_attachment_updates_post_meta() {
+		$webp = 'https://example.com/wp-content/uploads/2024/01/sample.webp';
+
+		$options = [
+			'logs' => true,
+		];
+
+		\WP_Mock::userFunction( 'is_wp_error' )
+			->once()
+			->with( $webp )
+			->andReturn( false );
+
+		\WP_Mock::userFunction( 'get_post_meta' )
+			->once()
+			->with( 1, 'icfw_img', true )
+			->andReturn( '' );
+
+		\WP_Mock::userFunction( 'update_post_meta' )
+			->once()
+			->with( 1, 'icfw_img', 'https://example.com/wp-content/uploads/2024/01/sample.webp' )
+			->andReturn( null );
+
+		$this->metadata->add_webp_meta_to_attachment( $webp, 1 );
+
+		$this->assertConditionsMet();
+	}
+}
